requirements.txt
Flask
celery[redis]
pyyaml
watchdog
Pillow
